Tests/LibWeb/Text/expected/WebAnimations/misc/animating-transform-mixed-units.txt

<!DOCTYPE html>
<style>
    div {
        position: absolute;
        width: 100px;
        height: 100px;
    }

    #el1 {
        animation: anim1 1s linear;
    }
    #el2 {
        animation: anim2 1s linear;
    }
    #el3 {
        animation: anim3 1s linear;
    }

    @keyframes anim1 {
        to {
            transform: translate(200%, 100px) scale(2)
        }
    }

    @keyframes anim2 {
        from {
            transform: translate(200%, 100px) scale(2)
        }
    }

    @keyframes anim3 {
        from {
            transform: translate(100%, 100px) scale(2)
        }
        to {
            transform: translate(200%, 300px) scale(4)
        }
    }
</style>
<script src="../../include.js"></script>
<div id="el1"></div>
<div id="el2"></div>
<div id="el3"></div>
<script>
    test(() => {
        const timeline = internals.createInternalAnimationTimeline();
        const divs = document.getElementsByTagName("div");
        for (const div of divs) {
            for (const anim of div.getAnimations())
                anim.timeline = timeline;
        }

        timeline.setTime(500);
        const [el1, el2, el3] = divs;

        println(`el1 transform: ${el1.style.transform}`);
    });
</script>
